Update and rename build.yml to deploy.yml
This adds a GitHub Pages deployment stage to the build workflow, which I thus renamed to “deploy.yml”

I’m not sure if you’re interested in adding this or not, but I figured it would be worth proposing the change anyway.

Note: I have _not_ actually tested this, because the repo I use for my own live DG site is private and therefore won’t let me use Pages for free. So if you decide to incorporate this change, please keep that in mind — there might be a missing parameter or a typo or something.